[#89]refactor: highlight.py 위치 변경
